Move options out of the helper package 🌿
