+++ pkg/options/delete.go
