+++ pkg/options/delete_test.go
